extract task creation functionality
